requirements.txt
pandas
numpy

scikit-learn
xgboost
lightgbm

mlflow

boto3

fastapi
uvicorn

python-dotenv
joblib

pytest
pyyaml